did code cleaning, and fixed a bug, now only min players can enter the game room, except for game 1(drawTheWord)
